4-10-25
- New: Added a Monolith Purification Kit. It can be used to turn gear into Monolith variants. Currently, defaults to certain outfits.
- Update: Cleric will now sell mod kits for Monolith.
- Update: Text files reworked to be more compatible and with edits from THIGHS applied where relevant. Also to not repeat "modification kit" as much.
- Update: BaS patches
The basic SVD (vanilla and BaS) was upgrading to the SVDS PMC. Corrected to upgrading from the actual SVDS.
The Monolith Purification Kit will now be involved in upgrading the different AK12 rifles.
- New: SKS Carbine in The Zone [Cr3pis]
Blocks the base>modern upgrade, as this mod makes them into explicitely different models.
With the patch for SWP in my Hisalute tweak patch, you can still cut down the older model into the SKS obrez, but not the Molot.
- New: Mosin Trio [Kmack]
Allows cutting the Kmack's Mosin duplicate and the carbine into the obrez version, and this to the Shiv.
With the patch for SWP in my Hisalute tweak patch, its combinations for its Mosin Obrez will be removed.